Added error catch for negative cross-sections
As cs_improved_1b.csv was found to have negative values for the cross-sections, I tested the effect of having negative weights on StatsBase, function sample. Its behaviour with negative weights is slightly odd. The position in the array and the values of the negative numbers affect the distribution of the output. Although the effect seems to be small and I predict it has not effect on the results of the simulation especially because the results were similar to cs_improved_1a.csv, most simulations will have to unfortunately be repeated.